+++ review/src/main/java/com/unavu/reviews/service/impl/ReviewServiceImpl.java
@@ -1,5 +1,6 @@
 package com.unavu.reviews.service.impl;
 
+import com.unavu.common.web.exception.ResourceActionNotAllowedException;
 import com.unavu.common.web.exception.ResourceAlreadyExistsException;
 import com.unavu.common.web.exception.ResourceNotFoundException;
 import com.unavu.reviews.dto.CreateReviewDto;
@@ -8,10 +9,10 @@ import com.unavu.reviews.dto.SearchReviewDto;
 import com.unavu.reviews.dto.UpdateReviewDto;
 import com.unavu.reviews.entity.Review;
 import com.unavu.reviews.mapper.ReviewMapper;
+import com.unavu.reviews.provider.CurrentUserProvider;
 import com.unavu.reviews.repository.ReviewRepository;
 import com.unavu.reviews.service.IReviewService;
 import com.unavu.reviews.service.client.RestaurantFeignClient;
-import com.unavu.reviews.service.client.UserFeignClient;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -19,6 +20,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Objects;
 import java.util.Optional;
 
 @Slf4j
@@ -27,8 +30,8 @@ import java.util.Optional;
 public class ReviewServiceImpl implements IReviewService {
 
     private final ReviewRepository reviewRepository;
-    private UserFeignClient userFeignClient;
     private RestaurantFeignClient restaurantFeignClient;
+    private CurrentUserProvider currentUserProvider;
     @Override
     public Page<ReviewDto> listReviews(Pageable pageable) {
         log.info("Fetching review list with pagination: page={}, size={}",
@@ -38,15 +41,15 @@ public class ReviewServiceImpl implements IReviewService {
 
     @Override
     public Page<ReviewDto> searchReviews(SearchReviewDto searchReviewDto, Pageable pageable) {
-        log.info("Searching reviews with filters: restaurantId={}, userId={}, rating={}, isRecommended={}",
+        log.info("Searching reviews with filters: restaurantId={}, reviewerId={}, rating={}, isRecommended={}",
         searchReviewDto.getRestaurantId(),
-        searchReviewDto.getUserId(),
+        searchReviewDto.getReviewerId(),
         searchReviewDto.getRating(),
         searchReviewDto.getIsRecommended());
 
         Specification<Review> spec=Specification
                 .where(ReviewSpecification.hasRestaurantId(searchReviewDto.getRestaurantId()))
-                .and(ReviewSpecification.hasUserId(searchReviewDto.getUserId()))
+                .and(ReviewSpecification.hasReviewerId(searchReviewDto.getReviewerId()))
                 .and(ReviewSpecification.hasRating(searchReviewDto.getRating()))
                 .and(ReviewSpecification.hasIsRecommended(searchReviewDto.getIsRecommended()));
 
@@ -62,9 +65,9 @@ public class ReviewServiceImpl implements IReviewService {
     }
 
     @Override
-    public Page<ReviewDto> getReviewsByUser(Long userId, Pageable pageable) {
-        log.info("Fetching review by userId={}", userId);
-        return reviewRepository.findByUserId(userId,pageable).map(ReviewMapper::toDto);
+    public Page<ReviewDto> getReviewsByUser(String keyCloakId, Pageable pageable) {
+        log.info("Fetching review by keyCloakId={}", keyCloakId);
+        return reviewRepository.findByReviewerId(keyCloakId,pageable).map(ReviewMapper::toDto);
     }
 
     @Override
@@ -83,27 +86,24 @@ public class ReviewServiceImpl implements IReviewService {
     @Override
     @Transactional
     public void createReview(CreateReviewDto createReviewDto) {
-        log.info("Creating review: restaurantId={}, userId={}, rating={}, title={}, comment={}, isRecommended={}",
+        log.info("Creating review: restaurantId={}, reviewerId={}, rating={}, title={}, comment={}, isRecommended={}",
                 createReviewDto.getRestaurantId(),
-                createReviewDto.getUserId(),
+                createReviewDto.getReviewerId(),
                 createReviewDto.getRating(),
                 createReviewDto.getTitle(),
                 createReviewDto.getComment(),
                 createReviewDto.getIsRecommended());
 
-        if (!userFeignClient.doesUserExist(createReviewDto.getUserId())) {
-            throw new ResourceNotFoundException("User","id", createReviewDto.getUserId());
-        }
         if (!restaurantFeignClient.doesRestaurantExist(createReviewDto.getRestaurantId())) {
             throw new ResourceNotFoundException("Restaurant","id", createReviewDto.getRestaurantId());
         }
 
         Optional<Review> optionalReview =
-                reviewRepository.findByUserIdAndRestaurantId(createReviewDto.getUserId(), createReviewDto.getRestaurantId());
+                reviewRepository.findByReviewerIdAndRestaurantId(createReviewDto.getReviewerId(), createReviewDto.getRestaurantId());
 
         if (optionalReview.isPresent()) {
-            log.warn("User has posted review already for restaurant: userId={}, restaurantId={}",
-                    createReviewDto.getUserId(),createReviewDto.getRestaurantId());
+            log.warn("User has posted review already for restaurant: reviewerId={}, restaurantId={}",
+                    createReviewDto.getReviewerId(),createReviewDto.getRestaurantId());
             throw new ResourceAlreadyExistsException(
                     "Review","restaurant",optionalReview
             );
@@ -125,7 +125,10 @@ public class ReviewServiceImpl implements IReviewService {
                     log.warn("Review not found for update, id={}", id);
                     return new ResourceNotFoundException("Review", "id", id.toString());
                 });
-
+        if(!Objects.equals(review.getReviewerId(), currentUserProvider.getCurrentUserId()))
+        {
+            throw new ResourceActionNotAllowedException("User cannot update other's review");
+        }
         ReviewMapper.updateEntity(updateReviewDto,review);
         reviewRepository.save(review);
 
@@ -142,7 +145,10 @@ public class ReviewServiceImpl implements IReviewService {
                     log.warn("Review not found for delete, id={}", id);
                     return new ResourceNotFoundException("Review", "id", id.toString());
                 });
-
+        if(!Objects.equals(review.getReviewerId(), currentUserProvider.getCurrentUserId()))
+        {
+            throw new ResourceActionNotAllowedException("User cannot update other's review");
+        }
         reviewRepository.delete(review);
         log.info("Review deleted successfully, id={}", id);
     }
